+++ android/src/main/java/com/freeraspreactnative/dispatchers/ExecutionStateDispatcher.kt
@@ -3,13 +3,15 @@ package com.freeraspreactnative.dispatchers
 import com.freeraspreactnative.events.RaspExecutionStateEvent
 import com.freeraspreactnative.interfaces.PluginExecutionStateListener
 
-internal class ExecutionStateDispatcher(private val listener: PluginExecutionStateListener) {
+internal object ExecutionStateDispatcher {
+  lateinit var listener: PluginExecutionStateListener
   private val cache = mutableSetOf<RaspExecutionStateEvent>()
 
   private var isAppInForeground = false
   private var isListenerRegistered = false
 
-  fun registerListener() {
+  fun registerListener(newListener: PluginExecutionStateListener) {
+    listener = newListener
     isListenerRegistered = true
     isAppInForeground = true
     flushCache()