+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.5.1] - 2026-03-16
+
+- Android SDK version: 18.0.4
+- iOS SDK version: 6.14.1
+
+### React Native
+
+#### Fixed
+
+- Fixed case where event dispatchers on Android were not initialized before calling onResume at the app launch
+
 ## [4.5.0] - 2026-03-03
 
 - Android SDK version: 18.0.4